+++ spec/investment_horse_racing/crawler/parser/entry_page_spec.rb
@@ -94,132 +94,16 @@ RSpec.describe InvestmentHorseRacing::Crawler::Parser::EntryPageParser do
     end
   end
 
-  describe "#valid?" do
-    context "2018-06-24 hanshin no 1 race entry page" do
-      it "is valid" do
-        expect(@parser).to be_valid
-      end
-    end
-
-    context "2018-06-24 hanshin no 11 race entry page" do
-      it "is valid" do
-        expect(@parser).to be_valid
-      end
-    end
-
-    context "error page" do
-      it "is invalid" do
-        expect(@parser_error).not_to be_valid
-      end
-    end
-  end
-
   describe "#related_links" do
     context "2018-06-24 hanshin no 1 race entry page" do
       it "is horse, jocky, trainer pages" do
-        expect(@parser.related_links).to contain_exactly(
-          "https://keiba.yahoo.co.jp/directory/horse/2015103590/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015104308/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015104979/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015102853/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015103335/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015104928/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015105363/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015100586/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015102694/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015100632/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015103557/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015104964/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015102837/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015103462/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015101618/",
-          "https://keiba.yahoo.co.jp/directory/horse/2015106259/",
-          "https://keiba.yahoo.co.jp/directory/jocky/05386/",
-          "https://keiba.yahoo.co.jp/directory/jocky/05339/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01116/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01018/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01165/",
-          "https://keiba.yahoo.co.jp/directory/jocky/00894/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01019/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01114/",
-          "https://keiba.yahoo.co.jp/directory/jocky/05203/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01088/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01154/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01014/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01126/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01130/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01166/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01034/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01157/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01120/",
-          "https://keiba.yahoo.co.jp/directory/trainer/00438/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01111/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01041/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01073/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01138/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01140/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01104/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01046/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01022/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01022/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01050/",
-          "https://keiba.yahoo.co.jp/directory/trainer/00356/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01066/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01078/")
+        expect(@parser.related_links).to be_nil
       end
     end
 
     context "2018-06-24 hanshin no 11 race entry page" do
       it "is horse, jockey, trainer pages" do
-        expect(@parser_11.related_links).to contain_exactly(
-          "https://keiba.yahoo.co.jp/directory/horse/2011103854/",
-          "https://keiba.yahoo.co.jp/directory/horse/2013102360/",
-          "https://keiba.yahoo.co.jp/directory/horse/2013106101/",
-          "https://keiba.yahoo.co.jp/directory/horse/2013106099/",
-          "https://keiba.yahoo.co.jp/directory/horse/2013110001/",
-          "https://keiba.yahoo.co.jp/directory/horse/2011104019/",
-          "https://keiba.yahoo.co.jp/directory/horse/2012104503/",
-          "https://keiba.yahoo.co.jp/directory/horse/2014106010/",
-          "https://keiba.yahoo.co.jp/directory/horse/2012104668/",
-          "https://keiba.yahoo.co.jp/directory/horse/2013106007/",
-          "https://keiba.yahoo.co.jp/directory/horse/2010101161/",
-          "https://keiba.yahoo.co.jp/directory/horse/2012106426/",
-          "https://keiba.yahoo.co.jp/directory/horse/2011190007/",
-          "https://keiba.yahoo.co.jp/directory/horse/2010102459/",
-          "https://keiba.yahoo.co.jp/directory/horse/2013105904/",
-          "https://keiba.yahoo.co.jp/directory/horse/2014101976/",
-          "https://keiba.yahoo.co.jp/directory/jocky/05203/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01130/",
-          "https://keiba.yahoo.co.jp/directory/jocky/05339/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01018/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01088/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01116/",
-          "https://keiba.yahoo.co.jp/directory/jocky/05386/",
-          "https://keiba.yahoo.co.jp/directory/jocky/00666/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01077/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01014/",
-          "https://keiba.yahoo.co.jp/directory/jocky/00894/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01019/",
-          "https://keiba.yahoo.co.jp/directory/jocky/05529/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01126/",
-          "https://keiba.yahoo.co.jp/directory/jocky/01032/",
-          "https://keiba.yahoo.co.jp/directory/jocky/05212/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01055/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01073/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01071/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01002/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01071/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01070/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01055/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01002/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01070/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01061/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01084/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01046/",
-          "https://keiba.yahoo.co.jp/directory/trainer/05576/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01058/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01126/",
-          "https://keiba.yahoo.co.jp/directory/trainer/01053/")
+        expect(@parser_11.related_links).to be_nil
       end
     end
   end